requirements.txt
asgiref==3.8.1
Django==4.2.19
django-appconf==1.1.0
django-environ==0.12.0
django-imagekit==5.0.0
pilkit==3.0
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
